- user id type set int

+++ src/main/kotlin/com/routing/MessagesRouting.kt
@@ -19,7 +19,7 @@ fun Application.configureMessageRouting(messagesDao: MessagesDao) {
         post("/$BASE/createMessage") {
             try {
                 val formParameters = call.receiveParameters()
-                val senderId = formParameters.getOrFail("senderId")
+                val senderId = formParameters.getOrFail("senderId").toInt()
                 val senderProfilePicUrl = formParameters.getOrFail("senderProfilePicUrl")
                 val roomId = formParameters.getOrFail<Int>("roomId")
                 val messageText = formParameters.getOrFail("messageText")
